Add support for varying endiannesses.
Currently only works with big and little endian. The base endianness -- that
is, the endianness of the output, is set by defining SZ_BASE_ENDIANNESS when
compiling snowball. The host endianness is set by defining SZ_ENDIANNESS.

If neither are specified, then the base and host endianness are both little
endian, because I haven't added a stage where the endianness is defined by a
configuration stage and there's no config file for snowball to load at the
moment.